Extract getPercentile to util folder
