+++ src/pages/bookmarks.tsx
@@ -11,19 +11,9 @@ import {
   useBookmarkIdsQuery,
   useTrendingProjectsQuery
 } from '@/graphql/generated/gql'
+import getPercentile from '@/util/getPercentile'
 
 // @TODO add category column to table
-
-const NUMERIC_FIELDS = [
-  'contributorCount',
-  'forkCount',
-  'issueCount',
-  'pullRequestCount',
-  'starCount'
-] as const
-
-type NumericField = (typeof NUMERIC_FIELDS)[number]
-type NumericFieldStats = Record<NumericField, number | null>
 
 /**
  * Project table with all bookmarks of a user
@@ -32,31 +22,6 @@ const Bookmarks = () => {
   const [data, setData] = useState<Project[]>([])
   const [filters, setFilters] = useState<ProjectFilter>({})
   const [sorting, setSorting] = useState<ProjectOrderBy | null>(defaultSort)
-
-  const getPercentileValue = (projects: Project[], percentile: number, sortDescending = true) => {
-    const result: NumericFieldStats = {
-      contributorCount: null,
-      forkCount: null,
-      issueCount: null,
-      pullRequestCount: null,
-      starCount: null
-    }
-
-    NUMERIC_FIELDS.forEach((field) => {
-      const sortedData = projects
-        .map((item) => item[field] ?? null)
-        .filter((item: number | null): item is number => item !== null)
-        .sort((a, b) => (sortDescending ? b - a : a - b))
-
-      const percentileIndex = Math.floor(sortedData.length * percentile)
-      result[field] =
-        percentileIndex < sortedData.length && sortedData.length > 0
-          ? sortedData[percentileIndex]
-          : null
-    })
-
-    return result
-  }
 
   const user = useUser()
 
@@ -98,10 +63,10 @@ const Bookmarks = () => {
       const projectData = urqlData?.projectCollection?.edges?.map((edge) => edge.node) as Project[]
       setData(projectData)
       setPercentileStats({
-        topTenPercent: getPercentileValue(projectData, 0.1),
-        bottomTenPercent: getPercentileValue(projectData, 0.1, false),
-        topTwentyPercent: getPercentileValue(projectData, 0.2),
-        bottomTwentyPercent: getPercentileValue(projectData, 0.2, false)
+        topTenPercent: getPercentile(projectData, 0.1),
+        bottomTenPercent: getPercentile(projectData, 0.1, false),
+        topTwentyPercent: getPercentile(projectData, 0.2),
+        bottomTwentyPercent: getPercentile(projectData, 0.2, false)
       })
     }
   }, [urqlData])
