+++ src/pages/index.tsx
@@ -10,47 +10,12 @@ import {
   useTrendingProjectsQuery
 } from '@/graphql/generated/gql'
 import defaultFilters from '@/components/page/overview/defaultFilters'
-
-const NUMERIC_FIELDS = [
-  'contributorCount',
-  'forkCount',
-  'issueCount',
-  'pullRequestCount',
-  'starCount'
-] as const
-
-type NumericField = (typeof NUMERIC_FIELDS)[number]
-type NumericFieldStats = Record<NumericField, number | null>
+import getPercentile from '@/util/getPercentile'
 
 const TrendingProjects = () => {
   const [data, setData] = useState<Project[]>([])
   const [filters, setFilters] = useState<ProjectFilter>(defaultFilters)
   const [sorting, setSorting] = useState<ProjectOrderBy | null>(defaultSort)
-
-  const getPercentileValue = (projects: Project[], percentile: number, sortDescending = true) => {
-    const result: NumericFieldStats = {
-      contributorCount: null,
-      forkCount: null,
-      issueCount: null,
-      pullRequestCount: null,
-      starCount: null
-    }
-
-    NUMERIC_FIELDS.forEach((field) => {
-      const sortedData = projects
-        .map((item) => item[field] ?? null)
-        .filter((item: number | null): item is number => item !== null)
-        .sort((a, b) => (sortDescending ? b - a : a - b))
-
-      const percentileIndex = Math.floor(sortedData.length * percentile)
-      result[field] =
-        percentileIndex < sortedData.length && sortedData.length > 0
-          ? sortedData[percentileIndex]
-          : null
-    })
-
-    return result
-  }
 
   const [percentileStats, setPercentileStats] = useState({
     topTenPercent: {},
@@ -77,10 +42,10 @@ const TrendingProjects = () => {
       const projectData = urqlData?.projectCollection?.edges?.map((edge) => edge.node) as Project[]
       setData(projectData)
       setPercentileStats({
-        topTenPercent: getPercentileValue(projectData, 0.1),
-        bottomTenPercent: getPercentileValue(projectData, 0.1, false),
-        topTwentyPercent: getPercentileValue(projectData, 0.2),
-        bottomTwentyPercent: getPercentileValue(projectData, 0.2, false)
+        topTenPercent: getPercentile(projectData, 0.1),
+        bottomTenPercent: getPercentile(projectData, 0.1, false),
+        topTwentyPercent: getPercentile(projectData, 0.2),
+        bottomTwentyPercent: getPercentile(projectData, 0.2, false)
       })
     }
   }, [urqlData])
